requirements.txt
# local package
-e .

# external requirements
absl-py==0.7.0
alabaster==0.7.12
astor==0.7.1
attrs==18.2.0
awscli==1.16.115
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
botocore==1.12.105
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cmake==3.13.3
colorama==0.3.9
coverage==4.5.2
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
dlib==19.5.1
docutils==0.14
entrypoints==0.3
face-recognition-models==0.3.0
flake8==3.7.7
gast==0.2.2
geojson==2.4.1
gmaps==0.8.4
google-streetview==1.2.4
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
kwconfig==1.1.7
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
opencv-python==4.0.0.21
packaging==19.0
pandocfilters==1.4.2
parso==0.3.4
pathlib2==2.3.3
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
pyasn1==0.4.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.0.0
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
tensorboard==1.13.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
